Reorganizes files
