+++ index.js
@@ -1,10 +1,10 @@
 const postcss = require('postcss');
 const Tidy = require('./Tidy');
-const getGlobalOptions = require('./lib/getGlobalOptions');
-const tidyShorthandProperty = require('./lib/tidy-shorthand-property');
-const tidyProperty = require('./lib/tidy-property');
-const tidyFunction = require('./lib/tidy-function');
-const tidyVar = require('./lib/tidy-var');
+const getGlobalOptions = require('./src/getGlobalOptions');
+const tidyShorthandProperty = require('./tidy-shorthand-property');
+const tidyProperty = require('./tidy-property');
+const tidyFunction = require('./tidy-function');
+const tidyVar = require('./tidy-var');
 
 /**
  * Parse rules and insert span and offset values.
