+++ lib/camelCaseString.js
