+++ lib/cleanClone.js
