+++ lib/compareStrings.js
