+++ lib/getObjectByProperty.js
