+++ lib/test/camelCaseString.test.js
@@ -1,4 +1,4 @@
-const camelCaseString = require('./camelCaseString');
+const camelCaseString = require('../camelCaseString');
 
 /**
  * Camelcase the hyphenated property names.
