+++ lib/test/cleanClone.test.js
@@ -1,7 +1,7 @@
 /* eslint-disable max-len */
-const { run } = require('../../test');
+const run = require('../../test');
 const postcss = require('postcss');
-const cleanClone = require('./cleanClone');
+const cleanClone = require('../cleanClone');
 
 /**
  * Create a test plugin for the `cleanClone` helper. Running a test plugin
