+++ lib/test/compareStrings.test.js
@@ -1,4 +1,4 @@
-const compareStrings = require('./compareStrings');
+const compareStrings = require('../compareStrings');
 
 /**
  * Compare numerical strings.
