+++ lib/test/getObjectByProperty.test.js
@@ -1,5 +1,5 @@
 /* eslint-disable max-len */
-const getObjectByProperty = require('./getObjectByProperty');
+const getObjectByProperty = require('../getObjectByProperty');
 
 /**
  * Find an object based on a given property value.
