+++ lib/test/parseAtruleParams.test.js
@@ -1,4 +1,4 @@
-const parseAtruleParams = require('./parseAtruleParams');
+const parseAtruleParams = require('../parseAtruleParams');
 
 /**
  * Parse AtRule params.
