+++ lib/test/sort.test.js
@@ -1,4 +1,4 @@
-const { strings, objectsByProperty } = require('./sort');
+const { strings, objectsByProperty } = require('../sort');
 
 /**
  * Sort array of strings.
