+++ lib/test/valuesHaveSameUnits.test.js
@@ -1,4 +1,4 @@
-const valuesHaveSameUnits = require('./valuesHaveSameUnits');
+const valuesHaveSameUnits = require('../valuesHaveSameUnits');
 
 /**
  * Check array values for like units.
