+++ lib/test/varPattern.test.js
@@ -1,4 +1,4 @@
-const varPattern = require('./varPattern');
+const varPattern = require('../varPattern');
 
 /**
  * Pattern to match CSS Custom Properties.
