+++ lib/valuesHaveSameUnits.js
