+++ lib/varPattern.js
