+++ src/getGlobalOptions.js
