+++ src/getLocalOptions.js
