+++ src/handleCustomProperties.js
