+++ test/index.js
@@ -14,20 +14,4 @@ const run = (input, output, opts, plugin = tidyColumns) => (
     })
 );
 
-/**
- * Options test runner.
- * Runs the plugin and verifies the options.
- */
-const runOptions = (input, output, opts, plugin = tidyColumns) => (
-  postcss([plugin(opts)])
-    .process(input, { from: undefined })
-    .then((result) => {
-      expect(result.options).toEqual(output);
-      expect(result.warnings().length).toBe(0);
-    })
-);
-
-module.exports = {
-  run,
-  runOptions,
-};
+module.exports = run;
